refactor: Infrastructure layer
