+++ PV260.Project/PV260.Project.Domain/Interfaces/Infrastructure/Email/IEmailSender.cs
@@ -1,6 +1,6 @@
 ﻿using PV260.Project.Domain.Models;
 
-namespace PV260.Project.Domain.Interfaces.Domain;
+namespace PV260.Project.Domain.Interfaces.Infrastructure.Email;
 
 public interface IEmailSender
 {
