+++ PV260.Project/PV260.Project.Domain/Interfaces/Infrastructure/Persistence/IUserRepository.cs
@@ -1,6 +1,6 @@
 ﻿using PV260.Project.Domain.Models;
 
-namespace PV260.Project.Domain.Interfaces.Infrastructure;
+namespace PV260.Project.Domain.Interfaces.Infrastructure.Persistence;
 
 public interface IUserRepository
 {
